requirements.txt
flask
Flask-Security-Too
Authlib
bcrypt
Flask-Migrate
Flask-SQLAlchemy
SQLAlchemy-Searchable
sqlalchemy-utils
Flask-WTF
email-validator

html2text
requests
gunicorn
python-dotenv
psycopg2-binary
beautifulsoup4
readabilipy

flask-dramatiq
redis
#dramatiq-pg

Flask-Mail

matplotlib
openai
scikit-learn